moved users examples out of users page
