+++ webapp/src/index.tsx
@@ -2,9 +2,9 @@ import { createRoot } from 'react-dom/client';
 import React from 'react';
 import './index.css';
 import reportWebVitals from './reportWebVitals';
-import { Chat, Landing, Play, Users } from './pages';
+import { Chat, Landing, Play, Users, Examples } from './pages';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
-import { CHAT_URL, PLAY_URL, USERS_URL } from './shared/urls';
+import { CHAT_URL, PLAY_URL, USERS_URL, EXAMPLES_URL } from './shared/urls';
 
 const container = document.getElementById('root');
 const root = createRoot(container!);
@@ -16,6 +16,7 @@ root.render(
         <Route path={USERS_URL} element={<Users />} />
         <Route path={PLAY_URL} element={<Play />} />
         <Route path={CHAT_URL} element={<Chat />} />
+        <Route path={EXAMPLES_URL} element={<Examples />} />
       </Routes>
     </BrowserRouter>
   </React.StrictMode>,
