+++ webapp/src/pages/Examples/Examples.css
@@ -1,0 +1,21 @@
+.examples {
+  margin: 0;
+  padding: 0;
+  min-height: 100vh;
+  background-color: var(--color-background);
+  color: var(--color-light);
+}
+
+.examples-wrapper {
+  width: 80%;
+  margin: 0 auto;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+  gap: 1rem;
+}
+
+.examples-nav-container {
+  width: 80%;
+}
