+++ webapp/src/pages/index.ts
@@ -2,3 +2,4 @@ export { default as Landing } from './Landing/Landing';
 export { default as Users } from './Users/Users';
 export { default as Play } from './Play/Play';
 export { default as Chat } from './Chat/Chat';
+export { default as Examples } from './Examples/Examples';
