+++ webapp/src/shared/urls.ts
@@ -1,3 +1,4 @@
 export const USERS_URL = '/users';
 export const PLAY_URL = '/play';
 export const CHAT_URL = '/chat';
+export const EXAMPLES_URL = '/examples';
